requirements.txt
Django==4.2.7
django-allauth==0.57.0
Pillow==10.1.0
exifread==3.0.0
numpy<2.0.0
shapely==2.0.2
django-storages==1.14.2
boto3==1.34.0
psycopg2-binary==2.9.9
python-decouple==3.8
geopy==2.4.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
